Import canvas source code
To make this project easier to build and contribute to, it is desirable
to import its source code into GitHub. The canvas sources use parts of
WireShark. Since the latter is licenses GPLv2 or later, this means that
canvas is also licensed GPLv2 or later.

Hence, I have determined that relicensing the whole code as GPLv3 (as
suggested by https://choosealicense.com/) is legal.